requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
attrs==24.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.3.2
dotenv==0.9.9
frozenlist==1.5.0
idna==3.7
multidict==6.1.0
pip==25.1
propcache==0.3.1
python-dotenv==1.1.0
requests==2.32.3
setuptools==78.1.1
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.4.0
wheel==0.45.1
yarl==1.18.0
psycopg2-binary
sqlalchemy